==> game_recommender/01_data_collection.ipynb <==
# Game Recommender — Compile Data

All data is extracted from [Backloggd](https://backloggd.com/games/lib/popular/) and [IGDB](https://www.igdb.com/).

# General
import re
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

# Web scraping
import requests
from bs4 import BeautifulSoup

# Utility
import time
import unidecode

import warnings
warnings.simplefilter('ignore')

Extract trending game titles, dev / genre tags and game summaries from Backloggd:

BASE_URL = 'https://backloggd.com/'
SEARCH_URL = "https://backloggd.com/games/lib/popular?page="

game_data = []

for page_number in tqdm(range(1, 41), total=40):
    response = requests.get(f"{SEARCH_URL}{page_number}")
    soup = BeautifulSoup(response.content, 'html.parser')

    try:
        # Fetch game titles along with href link to the game's info page
        for div in soup.find_all('div', {'class': 'col-2 my-2 px-1 px-md-2'}):
            title, game_info_url = div.div.text.strip(), div.a['href']
            game_data.append([title, f"{BASE_URL}{game_info_url}"])
    except Exception as e:
        print(f"Error at page number: {page_number} — {e}")

print(game_data[:5])

for i in tqdm(range(len(game_data)), total=len(game_data)):
    game_title, game_info_url = game_data[i][0], game_data[i][1]
    
    response = requests.get(game_info_url)
    soup = BeautifulSoup(response.content, 'html.parser')

    try:
        # Split dev tags and remove additional whitespace from each string
        developers = [
            re.sub('\s+', ' ', dev_name).strip() for dev_name in\
            soup.find('div', {'class': 'col-auto pl-lg-1 sub-title'})\
            .text.replace('by\n', '').strip().split(',')
        ]
        
        genre_tags = [tag.text for tag in soup.find_all('p', {'class': 'genre-tag'})]
        game_summary = soup.find('div', {'id': 'collapseSummary'}).text.strip()
    except Exception as e:
        print(f"Error at index pos; {i}, {game_title} – {e}")
        continue

    # Store data in new_games_df
    game_data[i].append(developers)
    game_data[i].append(genre_tags)
    game_data[i].append(game_summary)

# Sample row of data
game_data[0]

def reset_data(data: list) -> list:
    return [entry[:5] for entry in data]

game_data = reset_data(game_data)

Extract more tags from IGDB (platform, themes, storyline, game modes, keywords):

# API info for IGDB
API_URL = 'https://api.igdb.com/v4/games/'
CLIENT_ID = '7w2gvan50xfhdyjrjnshmx6o82tmtc'
CLIENT_SECRET = '4avly364nhovv2faj4d5l8pypeonmv'

# Get access token
post_url = 'https://id.twitch.tv/oauth2/token?'
params = {'client_id': CLIENT_ID, 'client_secret': CLIENT_SECRET, 'grant_type': 'client_credentials'}
response = requests.post(post_url, params=params)
access_token = response.json()['access_token']

for i in tqdm(range(len(game_data)), total=len(game_data)):
    game_title = unidecode.unidecode(game_data[i][0])
    game_query = f'search "{game_title}"; fields themes.*,game_modes.*,storyline,platforms.*,keywords.*,player_perspectives.*;'
    api_headers = {'headers': {'Client-ID': CLIENT_ID, 'Authorization': f'Bearer {access_token}'}, 'data': game_query}

    # Exception for pasring errors with API
    try:
        response = requests.post(API_URL, **api_headers)
        query_data = response.json()[0]
    except Exception as e:
        print(f"Error at index pos: {i}, {game_title} — {e}")

    for column_key in ['storyline', 'platforms', 'game_modes', 'themes', 'keywords', 'player_perspectives']:
        # Exception for missing data in query data
        try:
            if column_key == 'storyline':
                game_data[i].append(query_data[column_key])
            else:
                game_data[i].append([data['name'] for data in query_data[column_key]])
        except Exception as e:
            game_data[i].append([])

    time.sleep(0.5)

# Store game data in dataframe
df_columns = [
    'game_title', 'game_info_url', 'dev_team',
    'genre_tags', 'game_summary', 'storyline', 'platforms',
    'game_modes', 'themes', 'keywords', 'perspectives'
]
df = pd.DataFrame(data=game_data, columns=df_columns).drop(columns=['game_info_url'])
df.sample(3)

Save dataset:

df.to_csv('datasets/new/new_games_addon.csv', index=False)
