==> game_recommender/02_data_preprocessing.ipynb <==
# Game Recommender — Data Preprocessing

Notebook is designed to be executed once.

# General 
import re
import numpy as np
import pandas as pd
from tqdm.notebook import tqdm

# Utility
from collections import Counter
from ast import literal_eval

import warnings
warnings.simplefilter('ignore')

Load in data:

df = pd.read_csv('datasets/new/new_games_addon.csv')
print(df.shape, '\n')
df.sample(2)

Clean data:

df.info()

# Handle missing data
df[df.isna().any(axis=1)]

# Filter out rows with NaN values
df = df[~df.isna().any(axis=1)]

# Convert list like columns to Python lists
list_like_cols = ['dev_team', 'genre_tags', 'platforms', 'game_modes', 'themes', 'keywords', 'perspectives']
for col in list_like_cols:
    df[col] = df[col].apply(literal_eval)

df.sample(2)

# Handle duplicate titles / different editions of the same game

def remove_editions(game_title: str, dilimeter: str) -> bool:
    """
    Removes addtional editions of games already present in dataset.
    """
    if dilimeter in game_title and 'Edition' in game_title\
    or dilimeter in game_title and "Director's Cut" in game_title\
    or dilimeter in game_title and 'Deluxe' in game_title\
    or dilimeter in game_title and "Remastered" in game_title:
        title_prefix = game_title.split(dilimeter)[0]
        return title_prefix in set(df['game_title'])
    
    return False

df = df[~df['game_title'].apply(lambda x: remove_editions(x, ' - '))]
df = df[~df['game_title'].apply(lambda x: remove_editions(x, ':'))]
df = df[~df.duplicated(subset='game_summary')]  # Remove duplicate summaries
df = df[~df.duplicated(subset='game_title')]  # remove duplicate titles
df.shape

df = df.reset_index().drop(columns='index')
df.info()

Merge `genre_tags`, `themes`, `game_modes` and `keywords` into a single column:

# Since keywords contains the most tags, I'll be selecting the most common / what I think could be useful
tag_counts = Counter()
for index, row in df.iterrows():
    tag_counts.update(row['keywords'])

print('Top 20 keywords:', tag_counts.most_common(20))

# Identify more descriptive keywords for each game
key_tags = [
    '2d', 'metroidvania', 'anime', 'jrpg', 'post-apocalyptic', 'elves',
    'character customization', 'pvp', 'multiple endings', 'gothic', 'side-scrolling',
    'difficult', 'exploration', 'turn-based', 'assassin', 'immersive', 'medieval',
    'sexual content', 'sword & sorcery', 'female protagonist', 'wizards', 'aliens'
]

for index, row in tqdm(df.iterrows(), total=len(df)):
    # Append theme / game mode tags to genre tags
    row['genre_tags'] += row['themes']
    row['genre_tags'] += row['game_modes']

    # Concat game summary and storyline
    if row['storyline'][0] != '[':  # Cell not empty
        row['game_summary'] += ' ' + row['storyline']
    
    # Filter keyword tags and append to genre tags
    for tag in key_tags:
        if tag in set(row['keywords']) or tag in row['game_summary'].lower():  # <- Check if keyword tags appear in game summary
            if tag == 'jrpg' and 'RPG' in row['genre_tags']:  # JRPG and RPG are two different genres, with different gameplay mechanics
                row['genre_tags'].remove('RPG')
            
            row['genre_tags'].append(tag)

# Most common genre tags
def counts(tags: list):
    genre_counts.update(tags)

genre_counts = Counter()
df['genre_tags'].apply(counts)
most_common_genres = set([tag for tag, count in genre_counts.most_common(50)])
print(most_common_genres)

Since keyword tags can be incorrect I'll group games that come from a series with the same tags.

for index, row in tqdm(df.iterrows(), total=len(df)):
    all_game_titles = set(df[:index]['game_title'].tolist() + df[index+1:]['game_title'].tolist())
    game_series = []
    game_found = False
    
    # Identify games part of a series
    for game in all_game_titles:
        # Manual check for GTA / Elder Scrolls
        if 'Grand Theft Auto' in row['game_title'] and 'Grand Theft Auto' in game\
        or 'The Elder Scrolls' in row['game_title'] and 'The Elder Scrolls' in game\
        or row['game_title'] in game:
            game_found = True
            game_series.append(game)
    
    if game_found:
        game_series.append(row['game_title'])

    # Add new genre tags to each game in series
    series_df = df[df['game_title'].isin(game_series)]
    series_genre_tags = set()
    
    def get_genre_tags(genre_list: list): 
        for tag in genre_list:
            if tag in most_common_genres:
                series_genre_tags.add(tag)

    def overwrite_tags(genre_list: list, title: str) -> list:
        return list(series_genre_tags) if title in game_series else genre_list

    series_df['genre_tags'].apply(get_genre_tags)
    df['genre_tags'] = df.apply(lambda x: overwrite_tags(x['genre_tags'], x['game_title']), axis=1)

# Filter / re-order columns
df = df.drop(columns=['game_modes', 'themes', 'storyline'])
df = df.rename(columns={'genre_tags': 'primary_tags'})
df = df[['game_title', 'dev_team', 'platforms', 'primary_tags', 'keywords', 'game_summary']]
df.sample(2)

Save dataset:

df.to_csv('datasets/cleaned/new_games_CLEANED.csv', index=False)
